Add files via upload
This python file is used to extract the background noise from a long audio file. The code first loads the audio file and an excel/csv file containing the time stamp of the vehicles' presence in the audio file. The code then takes this information and by excluding these vehicles, crops the time between consecutive vehicles and save them in the destination folder.